Move statistics screens to new module
