+++ app/src/main/java/de/danoeh/antennapod/fragment/DownloadsFragment.java
@@ -19,6 +19,7 @@ import com.google.android.material.tabs.TabLayoutMediator;
 
 import de.danoeh.antennapod.R;
 import de.danoeh.antennapod.activity.MainActivity;
+import de.danoeh.antennapod.ui.common.PagedToolbarFragment;
 
 /**
  * Shows the CompletedDownloadsFragment and the RunningDownloadsFragment.
