+++ settings.gradle
@@ -17,3 +17,4 @@ include ':ui:app-start-intent'
 include ':ui:common'
 include ':ui:i18n'
 include ':ui:png-icons'
+include ':ui:statistics'
