+++ ui/common/src/main/res/layout/pager_fragment.xml
