+++ ui/statistics/README.md
@@ -1,0 +1,3 @@
+# :ui:statistics
+
+This module provides the statistics screens.
