+++ ui/statistics/src/main/res/layout/feed_statistics_dialog.xml
