+++ ui/statistics/src/main/res/layout/statistics_filter_dialog.xml
