+++ ui/statistics/src/main/res/layout/statistics_fragment.xml
