+++ ui/statistics/src/main/res/layout/statistics_listitem_total.xml
@@ -6,7 +6,7 @@
     android:layout_height="wrap_content"
     android:padding="16dp">
 
-    <de.danoeh.antennapod.view.PieChartView
+    <de.danoeh.antennapod.ui.statistics.PieChartView
         android:id="@+id/pie_chart"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
