+++ ui/statistics/src/main/res/layout/statistics_year_listitem.xml
